requirements.txt
torch
torchvision
numpy<2.0
pandas
scikit-learn
matplotlib
opencv-python
tqdm
Pillow
seaborn